requirements.txt
streamlit==1.32.0
openai==0.28.1
python-dotenv==1.0.0
Pillow==10.1.0
gTTS==2.5.0
streamlit-webrtc==0.47.1
pydub==0.25.1
